refactor: move addDiscoveredSolanaAccounts to separate function
