+++ app/scripts/metamask-controller.js
@@ -30,9 +30,6 @@ import {
 } from '@metamask/keyring-controller';
 import createFilterMiddleware from '@metamask/eth-json-rpc-filters';
 import createSubscriptionManager from '@metamask/eth-json-rpc-filters/subscriptionManager';
-///: BEGIN:ONLY_INCLUDE_IF(multi-srp)
-import { KeyringInternalSnapClient } from '@metamask/keyring-internal-snap-client';
-///: END:ONLY_INCLUDE_IF
 import { JsonRpcError, providerErrors, rpcErrors } from '@metamask/rpc-errors';
 
 import { Mutex } from 'await-semaphore';
@@ -149,7 +146,7 @@ import { isSnapId } from '@metamask/snaps-utils';
 
 import { Interface } from '@ethersproject/abi';
 import { abiERC1155, abiERC721 } from '@metamask/metamask-eth-abis';
-import { isEvmAccountType, SolAccountType, SolScope } from '@metamask/keyring-api';
+import { isEvmAccountType, SolAccountType } from '@metamask/keyring-api';
 import {
   hasProperty,
   hexToBigInt,
@@ -268,7 +265,7 @@ import { MultichainNetworks } from '../../shared/constants/multichain/networks';
 import { BRIDGE_API_BASE_URL } from '../../shared/constants/bridge';
 import { BridgeStatusAction } from '../../shared/types/bridge-status';
 ///: BEGIN:ONLY_INCLUDE_IF(multi-srp)
-import { SOLANA_WALLET_SNAP_ID } from '../../shared/lib/accounts';
+import { addDiscoveredSolanaAccounts } from '../../shared/lib/accounts';
 ///: END:ONLY_INCLUDE_IF
 import {
   ///: BEGIN:ONLY_INCLUDE_IF(build-mmi)
@@ -4817,19 +4814,31 @@ export default class MetamaskController extends EventEmitter {
     try {
       // Scan accounts until we find an empty one
       const chainId = this.#getGlobalChainId();
-      let entropySource = keyringId;
 
-      if (!entropySource) {
-        entropySource = this.keyringController.state.keyringsMetadata[0].id;
-      }
       const keyringSelector = keyringId
         ? { id: keyringId }
         : { type: KeyringTypes.hd };
 
-      const accounts = await this.keyringController.withKeyring(
+      const {
+        accounts,
+        ///: BEGIN:ONLY_INCLUDE_IF(multi-srp)
+        entropySource,
+        ///: END:ONLY_INCLUDE_IF(multi-srp)
+      } = await this.keyringController.withKeyring(
         keyringSelector,
-        async ({ keyring }) => {
-          return await keyring.getAccounts();
+        async ({
+          keyring,
+          ///: BEGIN:ONLY_INCLUDE_IF(multi-srp)
+          metadata,
+          ///: END:ONLY_INCLUDE_IF(multi-srp)
+        }) => {
+          const keyringAccounts = await keyring.getAccounts();
+          return {
+            accounts: keyringAccounts,
+            ///: BEGIN:ONLY_INCLUDE_IF(multi-srp)
+            entropySource: metadata.id,
+            ///: END:ONLY_INCLUDE_IF(multi-srp)
+          };
         },
       );
       let address = accounts[accounts.length - 1];
@@ -4870,42 +4879,14 @@ export default class MetamaskController extends EventEmitter {
           },
         );
       }
-
       ///: BEGIN:ONLY_INCLUDE_IF(multi-srp)
-      const client = new KeyringInternalSnapClient({
-        messenger: this.controllerMessenger,
-        snapId: SOLANA_WALLET_SNAP_ID,
-      });
-      // TODO: Use `withKeyring` instead.
       const keyring = await this.getSnapKeyring();
-      for (let index = 0; ; index++) {
-        const discovered = await client.discoverAccounts(
-          [SolScope.Mainnet, SolScope.Testnet, SolScope.Devnet],
-          entropySource,
-          index,
-        );
-
-        // We stop discovering accounts if none got discovered for that index.
-        if (discovered.length === 0) {
-          break;
-        }
-
-        await Promise.allSettled(
-          discovered.map(async (discoveredAccount) => {
-            const options = {
-              derivationPath: discoveredAccount.derivationPath,
-              entropySource,
-            };
-
-            await keyring.createAccount(SOLANA_WALLET_SNAP_ID, options, {
-              displayConfirmation: false,
-              displayAccountNameSuggestion: false,
-              setSelectedAccount: false,
-            });
-          }),
-        );
-      }
-      ///: END:ONLY_INCLUDE_IF
+      await addDiscoveredSolanaAccounts(
+        this.controllerMessenger,
+        entropySource,
+        keyring,
+      );
+      ///: END:ONLY_INCLUDE_IF(multi-srp)
     } catch (e) {
       log.warn(`Failed to add accounts with balance. Error: ${e}`);
     } finally {
